Moved algorithms headers into priv (crange.h, filter.h, utility.h). Include "stc/algorithm.h" instead.

+++ misc/benchmarks/various/binsearch_bench.c
@@ -36,6 +36,6 @@ int main(int argc, char const *argv[])
     }
     t = clock() - t;
     printf("elements: %d\n", (int)N);
-    printf("%s: found %lld %f\n", LABEL, count, (float)t/CLOCKS_PER_SEC);
+    printf("%s: found %d %f\n", LABEL, (int)count, (float)t/CLOCKS_PER_SEC);
     ivec_drop(&v);
 }
